Updating PieCharts
1. Changed GraphActivity to Java to reduce world hunger.
2. Added dependencies for MPAndroidCharts, a pieChart library.
3. Changed battery% to float, for reasons.
4. Added two pieCharts, to give us overall information about charging
and discharging of device over a Wi-Fi.
    a. If the decharged was charged more than it was discharged over a
    certain Wi-Fi AP, then we will see that AP entry in charged
    category.
    b. Similarly for discharged category.

+++ app/src/main/java/edu/osu/table/MyWorker.kt
@@ -69,7 +69,7 @@ class MyWorker(ctx: Context, params: WorkerParameters) : Worker(ctx, params) {
             }
 
             // Write Battery Percentage to Database
-            wirelessData.BatteryPerc = batteryPct!!.toDouble()
+            wirelessData.BatteryPerc = batteryPct!!.toFloat()
 
             // Final Insert to Database
             wirelessDao?.insert(wirelessData)
