+++ app/src/main/java/edu/osu/table/ui/WirelessData/WirelessData.kt
@@ -16,12 +16,12 @@ data class WirelessData(@PrimaryKey(autoGenerate = true) var id: Long?,
                         @ColumnInfo(name = "thoughput_Mbps") var ThroughputMpbs: Double,
                         //@ColumnInfo(name = "channel_bw") var ChBW: Double,
                         @ColumnInfo(name = "linkspeed") var LinkSpeed: Int,
-                        @ColumnInfo(name = "battery_percentage") var BatteryPerc: Double
+                        @ColumnInfo(name = "battery_percentage") var BatteryPerc: Float
 
 
 
 
 ){
-    constructor():this(null,0,  "", "",-1000, 0.0, -1,-1.0)
+    constructor():this(null,0,  "", "",-1000, 0.0, -1,-1.0f)
 
 }
